+++ packages/profile-sync-controller/src/controllers/authentication/AuthenticationController.ts
@@ -17,7 +17,12 @@ import {
   createSnapAllPublicKeysRequest,
   createSnapSignMessageRequest,
 } from './auth-snap-requests';
-import type { LoginResponse, SRPInterface, UserProfile } from '../../sdk';
+import type {
+  LoginResponse,
+  SRPInterface,
+  UserProfile,
+  UserProfileMetaMetrics,
+} from '../../sdk';
 import {
   assertMessageStartsWithMetamask,
   AuthType,
@@ -47,6 +52,10 @@ const metadata: StateMetadata<AuthenticationControllerState> = {
   },
 };
 
+type ControllerConfig = {
+  env: Env;
+};
+
 // Messenger Actions
 type CreateActionsObj<Controller extends keyof AuthenticationController> = {
   [K in Controller]: {
@@ -59,6 +68,7 @@ type ActionsObj = CreateActionsObj<
   | 'performSignOut'
   | 'getBearerToken'
   | 'getSessionProfile'
+  | 'getUserProfileMetaMetrics'
   | 'isSignedIn'
 >;
 export type Actions =
@@ -75,6 +85,8 @@ export type AuthenticationControllerGetBearerToken =
   ActionsObj['getBearerToken'];
 export type AuthenticationControllerGetSessionProfile =
   ActionsObj['getSessionProfile'];
+export type AuthenticationControllerGetUserProfileMetaMetrics =
+  ActionsObj['getUserProfileMetaMetrics'];
 export type AuthenticationControllerIsSignedIn = ActionsObj['isSignedIn'];
 
 export type AuthenticationControllerStateChangeEvent =
@@ -116,6 +128,10 @@ export default class AuthenticationController extends BaseController<
 
   readonly #auth: SRPInterface;
 
+  readonly #config: ControllerConfig = {
+    env: Env.PRD,
+  };
+
   #isUnlocked = false;
 
   readonly #keyringController = {
@@ -138,10 +154,12 @@ export default class AuthenticationController extends BaseController<
   constructor({
     messenger,
     state,
+    config,
     metametrics,
   }: {
     messenger: AuthenticationControllerMessenger;
     state?: AuthenticationControllerState;
+    config?: Partial<ControllerConfig>;
     /**
      * Not using the Messaging System as we
      * do not want to tie this strictly to extension
@@ -159,11 +177,16 @@ export default class AuthenticationController extends BaseController<
       throw new Error('`metametrics` field is required');
     }
 
+    this.#config = {
+      ...this.#config,
+      ...config,
+    };
+
     this.#metametrics = metametrics;
 
     this.#auth = new JwtBearerAuth(
       {
-        env: Env.PRD,
+        env: this.#config.env,
         platform: metametrics.agent,
         type: AuthType.SRP,
       },
@@ -212,6 +235,11 @@ export default class AuthenticationController extends BaseController<
     this.messagingSystem.registerActionHandler(
       'AuthenticationController:performSignOut',
       this.performSignOut.bind(this),
+    );
+
+    this.messagingSystem.registerActionHandler(
+      'AuthenticationController:getUserProfileMetaMetrics',
+      this.getUserProfileMetaMetrics.bind(this),
     );
   }
 
@@ -312,6 +340,11 @@ export default class AuthenticationController extends BaseController<
   ): Promise<UserProfile> {
     this.#assertIsUnlocked('getSessionProfile');
     return await this.#auth.getUserProfile(entropySourceId);
+  }
+
+  public async getUserProfileMetaMetrics(): Promise<UserProfileMetaMetrics> {
+    this.#assertIsUnlocked('getUserProfileMetaMetrics');
+    return await this.#auth.getUserProfileMetaMetrics();
   }
 
   public isSignedIn(): boolean {